Extract reusable functions for actix configuration
Allows someone to easily create their own Actix main with exograph
embedded along with other services. Separate functions for the
resolver and playground have been added.

Also added an EXO_SERVER_HOST env variable which will be used for
binding to if set.

See #883.
